重构demo页面结构
