+++ example/components/old-scroll/oldScroll.vue
@@ -1,0 +1,136 @@
+<template>
+  <div ref="wrapper">
+    <slot></slot>
+  </div>
+</template>
+
+<script type="text/ecmascript-6">
+  import BScroll from 'scroll/index'
+
+  const DIRECTION_H = 'horizontal'
+  const DIRECTION_V = 'vertical'
+  const COMPONENT_NAME = 'scroll'
+
+  export default {
+    name: COMPONENT_NAME,
+    props: {
+      probeType: {
+        type: Number,
+        default: 1
+      },
+      click: {
+        type: Boolean,
+        default: true
+      },
+      listenScroll: {
+        type: Boolean,
+        default: false
+      },
+      data: {
+        type: Array,
+        default: null
+      },
+      beforeScroll: {
+        type: Boolean,
+        default: false
+      },
+      refreshDelay: {
+        type: Number,
+        default: 20
+      },
+      direction: {
+        type: String,
+        default: DIRECTION_V
+      },
+      scrollbar: {
+        type: Object,
+        default: null
+      },
+      pullDownRefresh: {
+        type: Object,
+        default: null
+      },
+      pullUpLoad: {
+        type: Object,
+        default: null
+      }
+    },
+    mounted() {
+      setTimeout(() => {
+        this._initScroll()
+      }, 20)
+    },
+    methods: {
+      _initScroll() {
+        if (!this.$refs.wrapper) {
+          return
+        }
+        this.scroll = new BScroll(this.$refs.wrapper, {
+          probeType: this.probeType,
+          click: this.click,
+          scrollX: this.direction === DIRECTION_H,
+          scrollY: this.direction === DIRECTION_V,
+          scrollbar: this.scrollbar,
+          pullDownRefresh: this.pullDownRefresh,
+          pullUpLoad: this.pullUpLoad
+        })
+
+        if (this.listenScroll) {
+          this.scroll.on('scroll', (pos) => {
+            this.$emit('scroll', pos)
+          })
+        }
+
+        if (this.beforeScroll) {
+          this.scroll.on('beforeScrollStart', () => {
+            this.$emit('beforeScroll')
+          })
+        }
+
+        if (this.pullDownRefresh) {
+          this.scroll.on('pullingDown', () => {
+            this.$emit('pullingDown')
+          })
+        }
+
+        if (this.pullUpLoad) {
+          this.scroll.on('pullingUp', () => {
+            this.$emit('pullingUp')
+          })
+        }
+      },
+      disable() {
+        this.scroll && this.scroll.disable()
+      },
+      enable() {
+        this.scroll && this.scroll.enable()
+      },
+      refresh() {
+        this.scroll && this.scroll.refresh()
+      },
+      finishPullDown() {
+        this.scroll && this.scroll.finishPullDown()
+      },
+      finishPullUp() {
+        this.scroll && this.scroll.finishPullUp()
+      },
+      scrollTo() {
+        this.scroll && this.scroll.scrollTo.apply(this.scroll, arguments)
+      },
+      scrollToElement() {
+        this.scroll && this.scroll.scrollToElement.apply(this.scroll, arguments)
+      }
+    },
+    watch: {
+      data() {
+        setTimeout(() => {
+          this.refresh()
+        }, this.refreshDelay)
+      }
+    }
+  }
+</script>
+
+<style scoped lang="stylus" rel="stylesheet/stylus">
+
+</style>
