+++ example/components/page/page.vue
@@ -40,7 +40,7 @@
   }
 </script>
 
-<style lang="stylus" rel="stylesheet/stylus">
+<style scoped lang="stylus" rel="stylesheet/stylus">
   .page
     position: fixed
     z-index: 20
